Refactor debug logging
Add middleware so that debug logging can just be used throughout all
commands without any additional setup.
