+++ cmds/global.go
@@ -1,0 +1,5 @@
+// This package is just here to hold global variables since
+// it really didn't make sense to put them any other place.
+package cmds
+
+var debugMode bool
