+++ cmds/query.go
@@ -278,7 +278,6 @@ func getOutputFormat(c *cli.Context) string {
 }
 
 func queryRunCommand(c *cli.Context) error {
-	setDebugMode(c)
 	// Don't run if nothing is passed into the command
 	if len(os.Args) == 2 {
 		logAndDie("See --help for collins query usage")
